Java/src/main/java/problemsolving/leetcode/logic/containsDuplicateII/ContainsDuplicateII.java
package problemsolving.leetcode.logic.containsDuplicateII;

import java.util.HashMap;
import java.util.Map;

public class ContainsDuplicateII {
    public boolean containsNearbyDuplicate(int[] nums, int k) {
        Map<Integer, Boolean> map = new HashMap<>();
        for (int i = 0; i < nums.length; i++) {
            if (i > k) {
                map.remove(nums[i - k - 1]);
            }
            if (map.containsKey(nums[i])) {
                return true;
            }
            map.put(nums[i], true);
        }

        return false;
    }
}